requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scholarly==1.7.11
pymongo==4.6.1
elasticsearch==8.11.0
python-dotenv==1.0.0
pydantic==2.4.2
playwright==1.41.2 
requests==2.31.0
flask==3.0.0
apscheduler==3.10.4
pdfminer.six==20221105
openai==1.3.0
transformers==4.41.2
torch==2.3.0 